requirements.txt
streamlit==1.32.0
selenium==4.32.0
webdriver-manager==4.0.2
pandas==2.2.0
validate-email==1.3
py3-validate-email==1.0.5
tldextract==5.1.1
python-dotenv==1.0.1 